ai-service/requirements.txt
fastapi
uvicorn
python-multipart
rembg
pillow
numpy
